refactor(phase3d): optimizer / scheduler を library/optimizer.py に分離
train_util.py から以下を library/optimizer.py に移動:
- get_optimizer (AdamW / 8bit / Lion / DAdaptation / Prodigy / Adafactor /
  schedule-free / 任意の module.Class 形式に対応する統合ディスパッチャ)
- get_optimizer_train_eval_fn / is_schedulefree_optimizer / get_dummy_scheduler
- get_scheduler_fix (LR scheduler ファクトリ)
- append_lr_to_logs / append_lr_to_logs_with_names

get_my_scheduler は noise scheduler でサンプリング系のため対象外（PR-4 で
sampling.py 行きの予定）。

train_util.py は re-export shim を維持。tests/test_optimizer.py は明示 import
していたため、shim 非依存方針に従い from library.optimizer import に切替。

train_util.py: 1,969 → 1,379 行
library/optimizer.py: 637 行（新規）
累計 6,888 → 1,379 行（80% 削減、PR-3 完了）
